+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+**1.8**
+
+- Feature: Show Corruption Icons on gear, thanks SlipCore.
+- Feature: Better config layout, and allow saving settings per account.
+- Fix: Narcissus, Adibags, Bagnon, ElvUI, ArkInventory and LiteBag compatibility.
+- Cleanup.
+
 **1.7**
 
 - Feature: Show corruption level against item in character screen.